pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoeit"
version = "0.1.0"
description = "Automated Elicited Imitation Task (EIT) transcription pipeline for non-native Spanish speakers"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Tilak Gupta"},
]
keywords = ["speech", "transcription", "whisper", "EIT", "Spanish", "ASR"]

dependencies = [
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "noisereduce>=3.0.0",
    "scipy>=1.10.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
    "jiwer>=3.0.0",
    "tqdm>=4.65.0",
    "rapidfuzz>=3.0.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
mlx = ["mlx>=0.5.0", "mlx-whisper>=0.1.0"]
whispercpp = ["pywhispercpp>=1.0.0"]
faster-whisper = ["faster-whisper>=1.0.0"]
notebook = ["jupyter>=1.0.0", "ipykernel>=6.0.0", "matplotlib>=3.7.0", "ipywidgets>=8.0.0"]
all = ["mlx>=0.5.0", "mlx-whisper>=0.1.0", "pywhispercpp>=1.0.0", "faster-whisper>=1.0.0",
       "jupyter>=1.0.0", "ipykernel>=6.0.0", "matplotlib>=3.7.0", "ipywidgets>=8.0.0"]

[project.scripts]
autoeit = "autoeit.__main__:main"
eit-pipeline = "eit_pipeline.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
